requirements.txt
flet>=0.21.0
flask>=3.0.0
pydantic>=2.0.0
aiosqlite>=0.19.0
pypdf>=3.17.0
python-docx>=1.1.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
llama-cpp-python>=0.2.0
